being a bit more paranoid aboud the user's choice of line height

+++ calamari_ocr/ocr/dataset/imageprocessors/center_normalizer.py
@@ -31,9 +31,9 @@ class CenterNormalizer(ImageProcessor):
                                    borderType=cv.BORDER_CONSTANT)
         smoothed += .001 * cv.blur(smoothed, (w, int(h*.5)), borderType=cv.BORDER_CONSTANT)
 
-        a = np.argmax(smoothed, axis=0).astype(np.uint8)
+        a = np.argmax(smoothed, axis=0).astype(np.uint16)
         kernel = cv.getGaussianKernel(int((8.*h*self.extra)+1), h*self.extra)
-        center = cv.filter2D(a, cv.CV_8U, kernel, borderType=cv.BORDER_REFLECT).flatten()
+        center = cv.filter2D(a, cv.CV_16U, kernel, borderType=cv.BORDER_REFLECT).flatten()
 
         deltas = abs(np.arange(h)[:, np.newaxis] - center[np.newaxis, :])
         mad = np.mean(deltas[line != 0])
